DFC: react only to own template blocks, ignore levers everywhere, fix case pressure units
- Damage/repair reports now fire only for NBT template cells: air cells,
  foreign blocks placed inside the bounds and levers are ignored (both
  break and place); the fallback that reported any out-of-template block
  as STRUCTURE damage is removed
- Levers are dropped at template parse time: not required at assembly,
  not tracked by damage tracking, never produce mismatch fixes — any
  player wiring (levers anywhere) works without touching the structure
- ReactorStructure fallback validator no longer requires levers
- Case pressure was stored in kPa but printed as MPa (12 000.000 MPa on
  the sign): display/stats/DB now convert kPa → MPa consistently
- Case pressure target recomputed in kPa; broken-state jitter clamped
- Update the template test: 9 lever cells dropped, 981 cells checked

+++ ui-energy/src/main/java/com/ultimateimprovments/energy/generation/reactor/ReactorStatsDisplay.java
@@ -76,7 +76,7 @@ public final class ReactorStatsDisplay {
         player.sendMessage(MessageUtil.parse("<dark_gray>│ <gray>Integrity:   <white>" + reactor.getDisplayCoreShInt() + " %"));
         player.sendMessage(MessageUtil.parse("<dark_gray>│ <dark_aqua>═[ <aqua>Case data <dark_aqua>]═"));
         player.sendMessage(MessageUtil.parse("<dark_gray>│ <gray>Temperature: <white>" + reactor.getDisplayCoreCaseTemp() + " C*"));
-        player.sendMessage(MessageUtil.parse("<dark_gray>│ <gray>Pressure:    <white>" + String.format("%.3f", reactor.getDisplayCoreCasePress() / 1000.0) + " MPa"));
+        player.sendMessage(MessageUtil.parse("<dark_gray>│ <gray>Pressure:    <white>" + String.format("%.3f", (double) reactor.getDisplayCoreCasePress() / 1000.0) + " MPa"));
         player.sendMessage(MessageUtil.parse("<dark_gray>│ <gray>Integrity:   <white>" + reactor.getDisplayCoreCaseInt() + " %"));
         player.sendMessage(MessageUtil.parse("<dark_gray>│ <dark_purple>═[ <light_purple>Fusion data <dark_purple>]═"));
         player.sendMessage(MessageUtil.parse("<dark_gray>│ <gray>Speed:      <white>" + String.format("%.0f", reactor.getFusion().getSpeedPct()) + " %"));
